fixed some merge bugs
